updated script for .deb files with SD, including new sd.xpm

+++ src/scripts/make-deb.sh
@@ -6,7 +6,7 @@
 DEBLOG=deb.log.$$
 
 if [ -z $QT_STATIC ]; then
-    echo "QT_STATIC is not set. Please set it to the base directory of a statically compiled Qt";
+    echo "QT_STATIC is not set; to set it use -static for QT configuration. Please set it to the base directory of a statically compiled Qt";
     exit 1;
 fi
 
@@ -15,7 +15,7 @@ if [ -z $APP_VERSION ]; then echo "APP_VERSION is not set"; exit 1; fi
 #if [ -z $PREV_VERSION ]; then echo "PREV_VERSION is not set"; exit 1; fi
 
 if [ -z $HUSH_DIR ]; then
-    echo "HUSH_DIR is not set. Please set it to the base directory of hush3.git"
+    echo "HUSH_DIR is not set. Please set it to the src directory of hush3.git"
     exit 1;
 fi
 
@@ -34,7 +34,12 @@ if [ ! -f $HUSH_DIR/hush-tx ]; then
     exit 1;
 fi
 
-echo -n "Cleaning..............."
+if [ ! -f $HUSH_DIR/hush-smart-chain ]; then
+    echo "Couldn't find hush-smart-chain in $HUSH_DIR . Please build hush-smart-chain."
+    exit 1;
+fi
+
+echo "Cleaning..............."
 rm -rf bin/*
 rm -rf artifacts/*
 make distclean >/dev/null 2>&1
@@ -43,7 +48,7 @@ echo "[OK]"
 echo ""
 echo "[Building $APP_VERSION on" `lsb_release -r`" logging to $DEBLOG ]"
 
-echo -n "Translations............"
+echo "Translations............"
 QT_STATIC=$QT_STATIC bash src/scripts/dotranslations.sh >/dev/null
 echo -n "Configuring............"
 $QT_STATIC/bin/qmake silentdragon.pro -spec linux-clang CONFIG+=release > /dev/null
@@ -58,7 +63,7 @@ echo "[OK]"
 
 
 # Test for Qt
-echo -n "Static link............"
+echo "Static link............"
 if [[ $(ldd silentdragon | grep -i "Qt") ]]; then
     echo "FOUND QT; ABORT";
     exit 1
@@ -66,34 +71,51 @@ fi
 echo "[OK]"
 
 
-echo -n "Packaging.............."
+echo "Packaging.............."
 APP=SilentDragon-v$APP_VERSION
 DIR=bin/$APP
 mkdir $DIR > /dev/null
+
+#Organizing all bins & essentials to centralized folder for tar.gz
+echo "Organizing binaries & essentials.............."
+cp silentdragon $DIR > /dev/null
+cp $HUSH_DIR/hushd $DIR > /dev/null
+cp $HUSH_DIR/hush-cli $DIR > /dev/null
+cp $HUSH_DIR/hush-tx $DIR > /dev/null
+# cp $HUSH_DIR/hush-smart-chain $DIR > /dev/null
+
+cp $HUSH_DIR/../asmap.dat $DIR > /dev/null
+cp $HUSH_DIR/../sapling-spend.params $DIR > /dev/null
+cp $HUSH_DIR/../sapling-output.params $DIR > /dev/null
+cp README.md $DIR > /dev/null
+cp LICENSE $DIR > /dev/null
+
+echo "Stripping hush.............."
+cd $DIR
+strip hushd
+strip hush-cli
+strip hush-tx
+# strip hush-smart-chain
 strip silentdragon
+cd ../..
 
-cp silentdragon                  $DIR > /dev/null
-cp $HUSH_DIR/artifacts/hushd    $DIR > /dev/null
-cp $HUSH_DIR/artifacts/hush-cli $DIR > /dev/null
-cp $HUSH_DIR/artifacts/hush-tx $DIR > /dev/null
-cp README.md                      $DIR > /dev/null
-cp LICENSE                        $DIR > /dev/null
+echo "Compressing files.............."
+cd bin/
+tar -czf $APP.tar.gz ./$APP > /dev/null
 
-cd bin && tar czf $APP.tar.gz $DIR/ > /dev/null
-cd ..
-
+echo "Copy compressed file.............."
 mkdir artifacts >/dev/null 2>&1
-cp $DIR.tar.gz ./artifacts/$APP-linux.tar.gz
-echo "[OK]"
+cp $APP.tar.gz ./artifacts/$APP-linux.tar.gz
+echo -n "[OK]"
 
-
+echo "Verify Compressed File.............."
 if [ -f artifacts/$APP-linux.tar.gz ] ; then
     echo -n "Package contents......."
     # Test if the package is built OK
-    if tar tf "artifacts/$APP-linux.tar.gz" | wc -l | grep -q "9"; then
+    if tar -tf "artifacts/$APP-linux.tar.gz" | wc -l | grep -q "10"; then
         echo "[OK]"
     else
-        echo "[ERROR] Wrong number of files does not match 9"
+        echo "[ERROR] Wrong number of files does not match 11"
         exit 1
     fi
 else
@@ -101,25 +123,47 @@ else
     exit 1
 fi
 
-echo -n "Building deb..........."
-debdir=bin/deb/silentdragon-v$APP_VERSION
+echo "Building deb..........."
+debdir=deb/silentdragon-v$APP_VERSION
 mkdir -p $debdir > /dev/null
 mkdir    $debdir/DEBIAN
 mkdir -p $debdir/usr/local/bin
-
-cat src/scripts/control | sed "s/RELEASE_VERSION/$APP_VERSION/g" > $debdir/DEBIAN/control
-
-cp silentdragon                   $debdir/usr/local/bin/
-cp $HUSH_DIR/artifacts/hushd $debdir/usr/local/bin/hushd
-
+mkdir -p $debdir/usr/share/hush
+mkdir -p $debdir/usr/lib
 mkdir -p $debdir/usr/share/pixmaps/
-cp res/silentdragon.xpm           $debdir/usr/share/pixmaps/
+
+cat ../src/scripts/control | sed "s/RELEASE_VERSION/$APP_VERSION/g" > $debdir/DEBIAN/control
+
+echo "Copying bins..........."
+cp ../silentdragon $debdir/usr/local/bin/
+cp $HUSH_DIR/hushd $debdir/usr/local/bin/
+cp $HUSH_DIR/hush-cli $debdir/usr/local/bin/
+# cp $HUSH_DIR/hush-cli $debdir/usr/local/bin/
+# cp $HUSH_DIR/hush-smart-chain $debdir/usr/local/bin/
+
+echo "Copying core libraries from silentdragon binary..........."
+# copy the required shared libs to the target folder
+# create directories if required
+for lib in `ldd $debdir/usr/local/bin/silentdragon | cut -d'>' -f2 | awk '{print $1}'` ; do
+   if [ -f "$lib" ] ; then
+        cp -v "$lib" $debdir/usr/lib/
+   fi  
+done
+
+echo "Copying asmap..........."
+cp $HUSH_DIR/../asmap.dat $debdir/usr/share/hush/
+echo "Copying sapling params..........."
+cp $HUSH_DIR/../sapling-spend.params $debdir/usr/share/hush/
+cp $HUSH_DIR/../sapling-output.params $debdir/usr/share/hush/
+
+echo "Copying SilentDragon icon..........."
+cp ../res/silentdragon.xpm $debdir/usr/share/pixmaps/
 
 mkdir -p $debdir/usr/share/applications
-cp src/scripts/desktopentry    $debdir/usr/share/applications/silentdragon.desktop
+cp ../src/scripts/desktopentry $debdir/usr/share/applications/silentdragon.desktop
 
-dpkg-deb --build $debdir >/dev/null
-cp $debdir.deb                 artifacts/$DIR.deb
-echo "[OK]"
+dpkg-deb --build --root-owner-group $debdir >/dev/null
+
+echo "[Success! $APP .deb has been created in $APP/bin/deb]"
 
 exit 0
